hotfix(introduced-bugs): fix all the bugs that introduced in the previous release
