+++ package.json
@@ -39,16 +39,16 @@
     "@fontsource-variable/plus-jakarta-sans": "^5.1.0",
     "@lingui/macro": "^4.13.0",
     "@lingui/react": "^4.13.0",
-    "@mui/icons-material": "^6.1.4",
+    "@mui/icons-material": "^6.1.5",
     "@mui/lab": "^6.0.0-beta.12",
-    "@mui/material": "^6.1.4",
-    "@mui/system": "^6.1.4",
-    "@mui/x-charts": "^7.21.0",
-    "@mui/x-charts-pro": "^7.0.0-beta.5",
-    "@mui/x-data-grid-premium": "^7.21.0",
-    "@mui/x-date-pickers-pro": "^7.21.0",
-    "@mui/x-tree-view": "^7.21.0",
-    "@tanstack/react-query": "^5.59.15",
+    "@mui/material": "^6.1.5",
+    "@mui/system": "^6.1.5",
+    "@mui/x-charts": "^7.22.0",
+    "@mui/x-charts-pro": "^7.0.0-beta.6",
+    "@mui/x-data-grid-premium": "^7.22.0",
+    "@mui/x-date-pickers-pro": "^7.22.0",
+    "@mui/x-tree-view": "^7.22.0",
+    "@tanstack/react-query": "^5.59.16",
     "@uidotdev/usehooks": "^2.4.1",
     "axios": "^1.7.7",
     "d3": "^7.9.0",
@@ -61,7 +61,7 @@
     "js-cookie": "^3.0.5",
     "lowlight": "^3.1.0",
     "mustache": "^4.2.0",
-    "posthog-js": "^1.174.2",
+    "posthog-js": "^1.176.0",
     "qrcode": "^1.5.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
@@ -70,7 +70,7 @@
     "react-router-dom": "^6.27.0",
     "react-transition-group": "^4.4.5",
     "react-window": "^1.8.10",
-    "recharts": "^2.13.0",
+    "recharts": "~2.12.7",
     "trackjs": "^3.10.4",
     "typescript-parsec": "^0.3.4",
     "yaml": "^2.6.0"
@@ -95,20 +95,20 @@
     "@types/js-cookie": "^3.0.6",
     "@types/mustache": "^4.2.5",
     "@types/qrcode": "^1.5.5",
-    "@types/react": "^18.3.11",
+    "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@types/react-is": "^18.3.0",
     "@types/react-lazy-load-image-component": "^1.6.4",
     "@types/react-window": "^1.8.8",
     "@vitejs/plugin-react-swc": "^3.7.1",
-    "@vitest/coverage-v8": "^2.1.3",
-    "@vitest/ui": "^2.1.3",
+    "@vitest/coverage-v8": "^2.1.4",
+    "@vitest/ui": "^2.1.4",
     "babel-plugin-macros": "3.1.0",
     "esbuild": "^0.24.0",
     "eslint": "^9.13.0",
     "eslint-plugin-prettier": "^5.2.1",
     "eslint-plugin-react-hooks": "^5.0.0",
-    "eslint-plugin-react-refresh": "^0.4.13",
+    "eslint-plugin-react-refresh": "^0.4.14",
     "eslint-plugin-storybook": "^0.10.1",
     "globals": "^15.11.0",
     "husky": "^9.1.6",
@@ -122,15 +122,15 @@
     "tsc-files": "^1.1.4",
     "type-fest": "^4.26.1",
     "typescript": "^5.6.3",
-    "typescript-eslint": "^8.10.0",
+    "typescript-eslint": "^8.11.0",
     "unique-names-generator": "^4.7.1",
-    "uuid": "^10.0.0",
-    "vite": "^5.4.9",
+    "uuid": "^11.0.1",
+    "vite": "^5.4.10",
     "vite-plugin-html": "^3.2.2",
     "vite-plugin-mock-dev-server": "^1.8.0",
     "vite-plugin-svgr": "^4.2.0",
-    "vitest": "^2.1.3",
-    "web-vitals": "^4.2.3"
+    "vitest": "^2.1.4",
+    "web-vitals": "^4.2.4"
   },
   "browserslist": {
     "production": [
