+++ src/locales/en-US/messages.po
@@ -1107,7 +1107,7 @@ msgid "Define Roles"
 msgstr "Define Roles"
 
 #: src/pages/panel/shared/failed-checks/FailedCheckIgnoreButton.tsx:44
-#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:96
+#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:97
 msgid "Delayed Effect"
 msgstr "Delayed Effect"
 
@@ -1337,7 +1337,7 @@ msgstr "failed"
 msgid "Failed to retrieve changelog"
 msgstr "Failed to retrieve changelog"
 
-#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:200
+#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:201
 msgid "Failing benchmarks"
 msgstr "Failing benchmarks"
 
@@ -1468,7 +1468,7 @@ msgid "Hourly"
 msgstr "Hourly"
 
 #: src/pages/panel/benchmark-detail/BenchmarkDetailCheckDetail.tsx:116
-#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:111
+#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:112
 msgid "How to fix"
 msgstr "How to fix"
 
@@ -1532,7 +1532,7 @@ msgstr "Identifies storage volumes that are not attached to any running instance
 msgid "If this email address was used to create an account, instructions to reset your password will be sent to you. Please check your email."
 msgstr "If this email address was used to create an account, instructions to reset your password will be sent to you. Please check your email."
 
-#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:226
+#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:227
 msgid "If this security check does not apply to this resource, you can safely ignore it."
 msgstr "If this security check does not apply to this resource, you can safely ignore it."
 
@@ -1545,7 +1545,7 @@ msgstr "Ignore"
 msgid "Ignore all"
 msgstr "Ignore all"
 
-#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:87
+#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:88
 msgid "Ignored"
 msgstr "Ignored"
 
@@ -1816,7 +1816,7 @@ msgstr "month"
 msgid "Monthly"
 msgstr "Monthly"
 
-#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:251
+#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:252
 msgid "More info"
 msgstr "More info"
 
@@ -1946,7 +1946,7 @@ msgstr "No API token has been created."
 msgid "No Thanks"
 msgstr "No Thanks"
 
-#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:158
+#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:159
 msgid "Non-Compliant Resources"
 msgstr "Non-Compliant Resources"
 
@@ -2458,7 +2458,7 @@ msgstr "Review and assign"
 msgid "Review and assign (again)"
 msgstr "Review and assign (again)"
 
-#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:107
+#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:108
 msgid "Risk"
 msgstr "Risk"
 
@@ -2628,7 +2628,7 @@ msgstr "Severities"
 msgid "Severity"
 msgstr "Severity"
 
-#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:247
+#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:248
 msgid "Show resources"
 msgstr "Show resources"
 
@@ -2712,7 +2712,7 @@ msgid "Test"
 msgstr "Test"
 
 #: src/pages/panel/shared/failed-checks/FailedCheckIgnoreButton.tsx:50
-#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:100
+#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:101
 msgid "The change will be active from the next security scan onwards. Until the next scan, the resource will still show the failing check."
 msgstr "The change will be active from the next security scan onwards. Until the next scan, the resource will still show the failing check."
 
@@ -2815,7 +2815,7 @@ msgid "Token"
 msgstr "Token"
 
 #: src/pages/panel/security/OverallCard.tsx:123
-#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:117
+#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:118
 msgid "Top Non-Compliant Resources"
 msgstr "Top Non-Compliant Resources"
 
@@ -3171,7 +3171,7 @@ msgid "You need a payment method to change your product tier"
 msgstr "You need a payment method to change your product tier"
 
 #: src/pages/panel/shared/failed-checks/FailedCheckIgnoreButton.tsx:47
-#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:98
+#: src/pages/panel/shared/failed-checks/FailedChecks.tsx:99
 msgid "You've chosen to ignore this security check for the resource. Please note:"
 msgstr "You've chosen to ignore this security check for the resource. Please note:"
