+++ src/pages/panel/shared/utils/PieResourceCheckScore.tsx
@@ -1,5 +1,5 @@
 import { Box, ButtonBase, Stack, StackProps, styled, useTheme } from '@mui/material'
-import { MouseEventHandler, useMemo } from 'react'
+import { MouseEventHandler } from 'react'
 import { PieChart } from 'src/shared/charts'
 import { CircularScore } from 'src/shared/circular-score'
 import { shouldForwardPropWithBlackList } from 'src/shared/utils/shouldForwardProp'
@@ -44,9 +44,6 @@ export const PieResourceCheckScore = ({
   ...props
 }: PieResourceCheckScoreProps) => {
   const theme = useTheme()
-  const colors = useMemo(() => {
-    return { ...colorsBySeverity, Info: theme.palette.info.main }
-  }, [theme.palette.info.main])
   return (
     <Stack width={170} alignSelf="center" justifyContent="end" overflow="visible" flex={1} {...props}>
       <Box position="relative" height={170}>
@@ -58,14 +55,14 @@ export const PieResourceCheckScore = ({
             size={150}
             typographyProps={{
               variant: showPieChart ? 'h2' : 'h1',
-              sx: { transition: theme.transitions.create(['font-size']) },
+              sx: { transition: ({ transitions }) => transitions.create(['font-size']) },
             }}
           />
         </Stack>
         {showPieChart ? (
           <PieChartContainer opacity={hidingPieChart ? 0 : 1}>
             <PieChart
-              colors={colors}
+              colors={colorsBySeverity}
               data={data}
               showLabel
               pieProps={{
