+++ src/pages/panel/shared/utils/colorsBySeverity.ts
@@ -1,3 +1,4 @@
+import { colors } from '@mui/material'
 import { colorFromRedToGreen } from 'src/shared/constants'
 
 export const colorsBySeverity = {
@@ -5,6 +6,6 @@ export const colorsBySeverity = {
   High: colorFromRedToGreen[35],
   Medium: colorFromRedToGreen[50],
   Low: colorFromRedToGreen[65],
-  Info: 'grey.500',
+  Info: colors.grey[500],
   Passed: colorFromRedToGreen[100],
 } as const
